hardhat configuration updated for bscTestnet and hardhat tasks migrated to standalone hardhat scripts
