+++ package.json
@@ -19,7 +19,6 @@
   "author": "Venus",
   "license": "BSD-3-Clause",
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@nomiclabs/hardhat-etherscan": "2.1.3",
     "@nomiclabs/hardhat-truffle5": "^2.0.0",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
@@ -36,11 +35,13 @@
     "solparse": "^2.2.8"
   },
   "dependencies": {
+    "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@nomiclabs/hardhat-truffle5": "^2.0.3",
     "@nomiclabs/hardhat-web3": "^2.0.0",
+    "dotenv": "^14.2.0",
     "eth-saddle": "^0.1.21",
     "hardhat": "^2.8.0",
-    "dotenv": "^14.2.0"
+    "hardhat-deploy-ethers": "^0.3.0-beta.13"
   },
   "resolutions": {
     "scrypt.js": "https://registry.npmjs.org/@compound-finance/ethereumjs-wallet/-/ethereumjs-wallet-0.6.3.tgz",
